backend/pyproject.toml
[project]
name = "auto-fanban"
version = "0.1.0"
description = "DWG批处理系统 - 图签提取与文档生成"
requires-python = ">=3.13"
dependencies = [
    # Web 框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # 数据验证与配置
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    
    # CAD 处理
    "ezdxf>=1.3.0",
    
    # Office 文档处理
    "openpyxl>=3.1.5",
    "python-docx>=1.1.2",
    
    # PDF 处理
    "pypdf>=5.0.0",
    
    # 文件上传
    "python-multipart>=0.0.17",
    
    # Windows COM 自动化（仅Windows）
    "pywin32>=308; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "unit: 单元测试（快速，无外部依赖）",
    "integration: 集成测试（需要外部服务/文件）",
    "slow: 慢速测试（ODA转换/Office导出等）",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
